+++ docs/content/changelog.md
@@ -10,6 +10,8 @@ title = "Changelog"
   rename everything from `c4d.apex` to `c4ddev`
 - merge https://github.com/NiklasRosenstein/py-c4dtools into C4DDev
 - merge https://github.com/NiklasRosenstein/c4d-sublime-script into C4DDev
+- Rename SublimeScript to Scripting Server
+- The Scripting Server must be enabled explicitly from the C4DDev plugin menu
 
 ## v0.1.4